--- app/src/main/java/com/byeboo/app/data/dto/response/quest/QuestLikeResponseDto.kt
@@ -1,12 +1,0 @@
-package com.byeboo.app.data.dto.response.quest
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class QuestLikeResponseDto(
-    @SerialName("likeCount")
-    val likeCount: Int,
-    @SerialName("isLiked")
-    val isLiked: Boolean,
-)
