--- app/src/main/java/com/byeboo/app/domain/model/quest/QuestLikeUpdateModel.kt
@@ -1,7 +1,0 @@
-package com.byeboo.app.domain.model.quest
-
-data class QuestLikeUpdateModel(
-    val answerId: Long,
-    val heartCount: Int,
-    val isLiked: Boolean,
-)
